cleaned up the code base and removed the ipynb 2 src files for both implementations
